ui: stop assuming a header bar is ours, and type the signal ids
AdwHeaderBar is G_DECLARE_FINAL_TYPE, so GPasteUiHeader cannot be a subclass
of it and the buttons have nowhere to live but qdata -- which is also why
ADW_IS_HEADER_BAR() cannot tell our header bar from any other. Every accessor
asserted it anyway and then dereferenced whatever g_object_get_data()
returned, so passing a plain AdwHeaderBar crashed instead of warning. They
all go through one lookup now, which checks that the qdata is there and
leaves the caller with a warning rather than a NULL dereference.

The window's three "wait until the daemon connection is up" trampolines were
the same function three times over, each hand-rolling a two-slot gpointer
array to smuggle its argument through g_idle_add(); one deferred action
covers all three.

And a signal id is a gulong (or a guint, from g_signal_new), never a guint64:
nine arrays said otherwise.

+++ src/libgpaste/gpaste-daemon/gpaste-bus.c
@@ -27,7 +27,7 @@ enum
     LAST_SIGNAL
 };
 
-static guint64 signals[LAST_SIGNAL] = { 0 };
+static guint signals[LAST_SIGNAL] = { 0 };
 
 static void
 g_paste_bus_register_object (GPasteBus       *self,
