+++ src/ui/gpaste-ui-history.c
@@ -45,7 +45,7 @@ enum
     LAST_SIGNAL
 };
 
-static guint64 signals[LAST_SIGNAL] = { 0 };
+static guint signals[LAST_SIGNAL] = { 0 };
 
 enum
 {
